Refactor stix2.parsing into more focused modules: - stix2.registry, which contains the class mapping structure   and code for scanning stix2 modules for its initial population - stix2.registration, which contains code used to register custom   STIX types with the registry - stix2.parsing, which contains code for creating instances of   registered stix2 classes from raw dicts.
This is intended to reduce circular import problems, by giving
dependent code the ability to import a module which has exactly
the functionality it needs, without pulling a lot of other stuff
it doesn't need.  Fewer imports means less chance of an import
cycle.
