+++ stix2/test/v21/test_parsing.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 import pytest
 
 import stix2
-from stix2 import exceptions, parsing
+from stix2 import exceptions, parsing, registration, registry
 
 BUNDLE = {
     "type": "bundle",
@@ -78,11 +78,11 @@ def test_register_marking_with_version():
         _type = 'x-new-marking1'
         _properties = OrderedDict()
 
-    parsing._register_marking(NewMarking1, version='2.1')
+    registration._register_marking(NewMarking1, version='2.1')
     v = 'v21'
 
-    assert NewMarking1._type in parsing.STIX2_OBJ_MAPS[v]['markings']
-    assert v in str(parsing.STIX2_OBJ_MAPS[v]['markings'][NewMarking1._type])
+    assert NewMarking1._type in registry.STIX2_OBJ_MAPS[v]['markings']
+    assert v in str(registry.STIX2_OBJ_MAPS[v]['markings'][NewMarking1._type])
 
 
 @pytest.mark.xfail(reason="The default version is not 2.1", condition=stix2.DEFAULT_VERSION != "2.1")
@@ -92,8 +92,8 @@ def test_register_marking_with_no_version():
         _type = 'x-new-marking2'
         _properties = OrderedDict()
 
-    parsing._register_marking(NewMarking2)
+    registration._register_marking(NewMarking2)
     v = 'v21'
 
-    assert NewMarking2._type in parsing.STIX2_OBJ_MAPS[v]['markings']
-    assert v in str(parsing.STIX2_OBJ_MAPS[v]['markings'][NewMarking2._type])
+    assert NewMarking2._type in registry.STIX2_OBJ_MAPS[v]['markings']
+    assert v in str(registry.STIX2_OBJ_MAPS[v]['markings'][NewMarking2._type])
